prepare cart locking mechanism
