+++ src/Models/Cart.php
@@ -4,6 +4,7 @@ namespace Bazar\Models;
 
 use Bazar\Bazar;
 use Bazar\Concerns\Addressable;
+use Bazar\Concerns\InteractsWithDiscounts;
 use Bazar\Concerns\InteractsWithItems;
 use Bazar\Contracts\Discountable;
 use Bazar\Contracts\Itemable;
@@ -16,16 +17,7 @@ use Ramsey\Uuid\Uuid;
 
 class Cart extends Model implements Contract, Discountable, Itemable, Shippable
 {
-    use Addressable, InteractsWithItems;
-
-    /**
-     * The accessors to append to the model's array form.
-     *
-     * @var array
-     */
-    protected $appends = [
-        'currency',
-    ];
+    use Addressable, InteractsWithDiscounts, InteractsWithItems;
 
     /**
      * The attributes that should have default values.
@@ -34,6 +26,8 @@ class Cart extends Model implements Contract, Discountable, Itemable, Shippable
      */
     protected $attributes = [
         'discount' => 0,
+        'locked' => false,
+        'currency' => null,
     ];
 
     /**
@@ -42,6 +36,7 @@ class Cart extends Model implements Contract, Discountable, Itemable, Shippable
      * @var array
      */
     protected $casts = [
+        'locked' => 'bool',
         'discount' => 'float',
     ];
 
@@ -51,7 +46,9 @@ class Cart extends Model implements Contract, Discountable, Itemable, Shippable
      * @var array
      */
     protected $fillable = [
+        'locked',
         'discount',
+        'currency',
     ];
 
     /**
@@ -79,21 +76,58 @@ class Cart extends Model implements Contract, Discountable, Itemable, Shippable
     {
         static::creating(static function (Cart $cart): void {
             $cart->token = $cart->token ?: Uuid::uuid4();
+            $cart->currency = $cart->currency ?: Bazar::currency();
         });
     }
 
     /**
-     * Get the currency attribute.
+     * Lock the cart.
      *
-     * @return string
+     * @return void
      */
-    public function getCurrencyAttribute(): string
+    public function lock(): void
     {
-        return Bazar::currency();
+        if (! $this->locked) {
+            $this->fill(['locked' => true])->save();
+        }
     }
 
     /**
-     * Scope a query to only include expired carts.
+     * Unlock the cart.
+     *
+     * @return void
+     */
+    public function unlock(): void
+    {
+        if ($this->locked) {
+            $this->fill(['locked' => false])->save();
+        }
+    }
+
+    /**
+     * Scope a query to only include the locked carts.
+     *
+     * @param  \Illuminate\Database\Eloquent\Builder  $query
+     * @return \Illuminate\Database\Eloquent\Builder
+     */
+    public function scopeLocked(Builder $query): Builder
+    {
+        return $query->where('locked', true);
+    }
+
+    /**
+     * Scope a query to only include the unlocked carts.
+     *
+     * @param  \Illuminate\Database\Eloquent\Builder  $query
+     * @return \Illuminate\Database\Eloquent\Builder
+     */
+    public function scopeUnlocked(Builder $query): Builder
+    {
+        return $query->where('locked', false);
+    }
+
+    /**
+     * Scope a query to only include the expired carts.
      *
      * @param  \Illuminate\Database\Eloquent\Builder  $query
      * @return \Illuminate\Database\Eloquent\Builder
