+++ src/Models/Order.php
@@ -6,6 +6,7 @@ use Bazar\Bazar;
 use Bazar\Concerns\Addressable;
 use Bazar\Concerns\BazarRoutable;
 use Bazar\Concerns\Filterable;
+use Bazar\Concerns\InteractsWithDiscounts;
 use Bazar\Concerns\InteractsWithItems;
 use Bazar\Contracts\Breadcrumbable;
 use Bazar\Contracts\Discountable;
@@ -23,7 +24,7 @@ use Illuminate\Support\Collection;
 
 class Order extends Model implements Breadcrumbable, Contract, Discountable, Itemable, Shippable
 {
-    use Addressable, BazarRoutable, Filterable, InteractsWithItems, SoftDeletes;
+    use Addressable, BazarRoutable, Filterable, InteractsWithDiscounts, InteractsWithItems, SoftDeletes;
 
     /**
      * The accessors to append to the model's array form.
@@ -148,21 +149,6 @@ class Order extends Model implements Breadcrumbable, Contract, Discountable, Ite
     public function transactions(): HasMany
     {
         return $this->hasMany(TransactionProxy::getProxiedClass());
-    }
-
-    /**
-     * Get the currency attribute.
-     *
-     * @param  string|null  $value
-     * @return string
-     */
-    public function getCurrencyAttribute(string $value = null): string
-    {
-        if (! is_null($value) && in_array($value, array_keys(Bazar::currencies()))) {
-            return $value;
-        }
-
-        return Bazar::currency();
     }
 
     /**
